add: added new header-file with type-definitions (def.h) + additional callback
The new callback is needed because bindgen cannot infer type-casted macros. For example the macro '#define SCIP_INVALID (double)1e99' is too complicated. This is because of the type-cast to double. If we simply remove the (double), then bindgen CAN generate a binding for it. It also defaults to an `f64` in Rust, which is what the C-representation would look like too. The callback should only be used for types of which we know that Rust can infer the datatype. Therefore, I made sure that it only goes off on;
1. macros that the user deliberately targets (added via `target`-method)
2. only looks at casts to clang-keywords. So no weird custom mumbo-jumbo casts